+++ src/test/scala/uk/gov/hmrc/test/ui/pages2/Journeys.scala
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.test.ui.pages2
+
+import uk.gov.hmrc.test.ui.cucumber.stepdefs.World
+import uk.gov.hmrc.test.ui.pages2.application.ApplicationDetailsPage
+
+object Journeys extends Robot with World {
+
+  def openStartPage(): ServiceStartPage = {
+    navigateTo("")
+    ServiceStartPage()
+  }
+
+  def signIn(): DashboardPage = {
+    signInViaLdap()
+  }
+
+  def signInViaLdap(): DashboardPage = {
+    Journeys
+      .openStartPage()
+      .startNow()
+      .signInViaLdap()
+      .signInWithDefaults()
+  }
+
+  def signInViaStride(): DashboardPage = {
+    Journeys
+      .openStartPage()
+      .startNow()
+      .signInViaStride()
+      .signIn()
+  }
+
+  def registerAnApplication(dashboardPage: DashboardPage): ApplicationDetailsPage = {
+    dashboardPage
+      .registerAnApplication()
+      .setApplicationName(application.name)
+      .doNotAddTeamMembers()
+      .registerApplication()
+      .viewRegisteredApplication()
+  }
+
+  def signInAndRegisterAnApplication(): ApplicationDetailsPage = {
+    signIn().registerAnApplicationJourney()
+  }
+
+}
